+++ par2/par2repairer.cpp
@@ -379,7 +379,7 @@ bool Par2Repairer::LoadPacketsFromFile(string filename)
   // Did we actually find any interesting packets
   if (packets > 0)
   {
-    PDEBUG(DEBUG_MIN, "par2file %s: %i packets (%i recovery)",filename.c_str(), packets, recoverypackets);
+    PDEBUG(DEBUG_MIN, "par2file %s: %i packets (%i recovery)",filename.c_str(), (int)packets, (int)recoverypackets);
     /*cout << "Loaded " << packets << " new packets";
     if (recoverypackets > 0) cout << " including " << recoverypackets << " recovery blocks";
     cout << endl;*/
